+++ shared/DSPFilters/source/Biquad.cpp
@@ -122,18 +122,18 @@ void BiquadBase::setCoefficients (double a0, double a1, double a2,
   assert (!Dsp::is_nan (a0) && !Dsp::is_nan (a1) && !Dsp::is_nan (a2) &&
           !Dsp::is_nan (b0) && !Dsp::is_nan (b1) && !Dsp::is_nan (b2));
 
-  m_a0 = a0;
-  m_a1 = a1/a0;
-  m_a2 = a2/a0;
-  m_b0 = b0/a0;
-  m_b1 = b1/a0;
-  m_b2 = b2/a0;
+  m_a0 = (float)a0;
+  m_a1 = (float)(a1/a0);
+  m_a2 = (float)(a2/a0);
+  m_b0 = (float)(b0/a0);
+  m_b1 = (float)(b1/a0);
+  m_b2 = (float)(b2/a0);
 #ifdef __SSE__
     m_va0 = _mm_set1_ps(m_a0);
     m_vb0 = _mm_set1_ps(m_b0);
     m_vab12 = _mm_set_ps(m_b2, m_b1, m_a2, m_a1);
 #elif defined(__ARM_NEON__)
-    m_va0 = vdup_n_f32(m_a0);
+  m_va0 = vdup_n_f32(m_a0);
   m_vb0 = vdup_n_f32(m_b0);
   vsetq_lane_f32(m_a1, m_vab12, 0);
   vsetq_lane_f32(m_b1, m_vab12, 1);
